requirements.txt
PyQt6
PyQt6-QScintilla
Pillow
PyWavefront
PyOpenGL
numpy
